requirements.txt
yt-dlp==2024.8.6
pillow==10.4.0